requirements.txt
torch
torchvision
torchaudio
datasets
ipywidgets
transformers
numba
openai-whisper